requirements.txt
streamlit==1.35.0
torch==2.2.2
torchvision==0.17.2
pillow==10.3.0
huggingface_hub==0.23.5
numpy==1.26.4
opencv-python==4.9.0.80